Move CLI out of `command/CLI` folder. (#90)
