+++ engine/src/command/CLI.cpp
@@ -1,4 +1,4 @@
-#include "command/CLI/CLI.h"
+#include "command/CLI.h"
 #include "Engine.h"
 #include "board/Board.h"
 #include "debug/perft.h"
